refactoring, moving faster_transmission into seir_abm & renaming to model.py
